requirements.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==23.2.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
gunicorn==22.0.0
h11==0.14.0
jax==0.4.26
kiwisolver==1.4.5
matplotlib==3.8.4
mediapipe==0.10.11
ml-dtypes==0.4.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.1
protobuf==3.20.3
psycopg2-binary==2.9.9
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
virtualenv==20.26.1
whitenoise==6.6.0
